Completed Todo app

+++ src/todos/todos.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { GetTodoFilterDto } from './dto/get-todo-filter.dto';
 import { TodoStatus } from './todo-status.enum';
@@ -13,8 +14,8 @@ export class TodosService {
         private todoRepository : TodoRepository
     ) {}
 
-    async getTodoById(id: number) : Promise<Todo>{
-        const found = await this.todoRepository.findOne(id);
+    async getTodoById(id: number,user: User) : Promise<Todo>{
+        const found = await this.todoRepository.findOne({ where: {id, userId: user.id}});
         if(!found) {
                   throw new NotFoundException('Todo With Id $id Not found');
         }
@@ -22,27 +23,29 @@ export class TodosService {
 
     }
 
-    async createTodos(createTodoDto: CreateTodoDto) : Promise<Todo> {
-      return this.todoRepository.createTodo(createTodoDto)
+    async createTodos(createTodoDto: CreateTodoDto,
+        user: User
+        ) : Promise<Todo> {
+      return this.todoRepository.createTodo(createTodoDto, user)
     }
 
-    async deleteTodo(id: number) : Promise<void> {
-       const result = await  this.todoRepository.delete(id)
+    async deleteTodo(id: number,user: User) : Promise<void> {
+       const result = await  this.todoRepository.delete({id, userId: user.id})
        if(result.affected === 0){
         throw new NotFoundException('Todo With Id $id Not found');
        }
     }
 
-     async updateTodoStatus(id: number,  status: TodoStatus): Promise<Todo> {
-        const todo = await this.getTodoById(id)
+     async updateTodoStatus(id: number,  status: TodoStatus, user: User): Promise<Todo> {
+        const todo = await this.getTodoById(id, user)
         todo.status = status;
         await todo.save();
         return todo;
 
     }
 
-    async getTasks(filterDto : GetTodoFilterDto) : Promise<Todo[]>{
-        return this.todoRepository.getTodos(filterDto)
+    async getTodos(filterDto : GetTodoFilterDto, user: User) : Promise<Todo[]>{
+        return this.todoRepository.getTodos(filterDto, user)
 
     }
